+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Added:**
 - Network Service - Added Network Security Rule.
 - Network Service - Added Get Available IP Addresses count in Subnet
+- Compute Service - Virtual Machine Extension
 
 ## 0.0.6